\name{BEACH-package}
\alias{BEACH-package}
\alias{BEACH}
\docType{package}
\title{
Biometric Exploratory Analysis Creation House
\packageTitle{BEACH}
}
\description{
A platform is provided for interactive analyses with a goal of totally easy to develop, deploy, interact, and explore (TEDDIE). Using this package, users can create customized analyses and make them available to end users who can perform interactive analyses and save analyses to RTF or HTML files. It allows developers to focus on R code for analysis, instead of dealing with html or shiny code.
\packageDescription{BEACH}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BEACH}
\packageIndices{BEACH}

}
\author{
Danni Yu and Michael Man
\packageAuthor{BEACH}

Maintainer: \packageMaintainer{BEACH} Danni Yu <danni.yu@gmail.com>
}
%%\references{
%%~~ Literature or other references for background information ~~
%%}
\keyword{ BEACH }
%%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}
\examples{
#  runBEACH()
}
